Improve player teleportation in wilderness
Added logic to the WildernessCommand class to make sure the teleport locations are appropriate. The code now ensures players are not teleported to invalid spots (such as in water, lava or portals) by checking the material of the surrounding blocks. A retry mechanism has also been added for finding valid locations.