Refactor FarmWorldManager and improve logging
Made use of the forEach method for iterating over retrieved FarmWorlds to enhance readability. Introduced logging to inform when no FarmWorld is found in the database or when a FarmWorld is deleted. Notably, added a comment to explain the conditional check for deleting a world only when its file no longer exists on the system. Optimized the "expired lifetime" check to ensure the deletion of FarmWorld only when it's due.